ISSUE-12 - Allow the writing of a stream to a file.
Added a new method to the file manager and adapters to write a stream to a file.

Implemented appropriate tests around that, including verifying that current AWS behavior remains the same.

Renamed some tests to remove 'async' from their class names.

Alphabeticised the functions under FileManager and S3Adapter.Files more.
