+++ src/Baseline.Filesystem/Responses/File/WriteStreamToFileResponse.cs
@@ -1,0 +1,10 @@
+namespace Baseline.Filesystem
+{
+    /// <summary>
+    /// Response returned from the WriteStreamAsync method that contains information about whether or not the operation
+    /// was a success and the details about what was created.
+    /// </summary>
+    public class WriteStreamToFileResponse
+    {
+    }
+}
