+++ test/Baseline.Filesystem.Tests/FileManagerTests/GetTests.cs
@@ -7,29 +7,29 @@ using Xunit;
 
 namespace Baseline.Filesystem.Tests.FileManagerTests
 {
-    public class TouchAsyncTests : BaseManagerUsageTest
+    public class GetTests : BaseManagerUsageTest
     {
         [Fact]
         public async Task It_Throws_An_Exception_If_The_Requested_Adapter_Name_Is_Not_Registered()
         {
-            Func<Task> func = async () => await FileManager.TouchAsync(
-                new TouchFileRequest { FilePath = "a".AsBaselineFilesystemPath() },
+            Func<Task> func = async () => await FileManager.GetAsync(
+                new GetFileRequest { FilePath = "a".AsBaselineFilesystemPath() },
                 "foo"
             );
             await func.Should().ThrowAsync<AdapterNotFoundException>();
         }
 
         [Fact]
-        public async Task It_Throws_An_Exception_If_The_TouchFileRequest_Was_Null()
+        public async Task It_Throws_An_Exception_If_The_Request_Was_Null()
         {
-            Func<Task> func = async () => await FileManager.TouchAsync(null);
+            Func<Task> func = async () => await FileManager.GetAsync(null);
             await func.Should().ThrowExactlyAsync<ArgumentNullException>();
         }
 
         [Fact]
-        public async Task It_Throws_An_Exception_If_The_Path_For_The_TouchFileRequest_Was_Null()
+        public async Task It_Throws_An_Exception_If_The_Path_For_The_Request_Was_Null()
         {
-            Func<Task> func = async () => await FileManager.TouchAsync(new TouchFileRequest());
+            Func<Task> func = async () => await FileManager.GetAsync(new GetFileRequest());
             await func.Should().ThrowExactlyAsync<ArgumentNullException>();
         }
 
@@ -38,19 +38,19 @@ namespace Baseline.Filesystem.Tests.FileManagerTests
         {
             var path = "/users/Foo/bar/Destiny/XYZ/BARTINO/".AsBaselineFilesystemPath();
             
-            Func<Task> func = async () => await FileManager.TouchAsync(new TouchFileRequest { FilePath = path });
+            Func<Task> func = async () => await FileManager.GetAsync(new GetFileRequest { FilePath = path });
             await func.Should().ThrowExactlyAsync<PathIsADirectoryException>();
         }
         
         [Fact]
-        public async Task It_Invokes_The_Matching_Adapters_Touch_File_Method_And_Wraps_The_Response()
+        public async Task It_Invokes_The_Matching_Adapters_Get_File_Method_And_Wraps_The_Response()
         {
             Adapter
-                .Setup(x => x.TouchFileAsync(It.IsAny<TouchFileRequest>(), It.IsAny<CancellationToken>()))
+                .Setup(x => x.GetFileAsync(It.IsAny<GetFileRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(new FileRepresentation { Path = new PathRepresentation() })
                 .Verifiable();
             
-            var response = await FileManager.TouchAsync(new TouchFileRequest { FilePath = "a".AsBaselineFilesystemPath() });
+            var response = await FileManager.GetAsync(new GetFileRequest { FilePath = "a".AsBaselineFilesystemPath() });
             response.AdapterName.Should().Be("default");
             
             Adapter.VerifyAll();
