+++ test/Baseline.Filesystem.Tests/FileManagerTests/ReadAsStringTests.cs
@@ -7,13 +7,13 @@ using Xunit;
 
 namespace Baseline.Filesystem.Tests.FileManagerTests
 {
-    public class DeleteAsyncTests : BaseManagerUsageTest
+    public class ReadAsStringTests : BaseManagerUsageTest
     {
         [Fact]
         public async Task It_Throws_An_Exception_If_The_Requested_Adapter_Name_Is_Not_Registered()
         {
-            Func<Task> func = async () => await FileManager.DeleteAsync(
-                new DeleteFileRequest { FilePath = "a".AsBaselineFilesystemPath() },
+            Func<Task> func = async () => await FileManager.ReadAsStringAsync(
+                new ReadFileAsStringRequest { FilePath = "a".AsBaselineFilesystemPath() },
                 "foo"
             );
             await func.Should().ThrowAsync<AdapterNotFoundException>();
@@ -22,14 +22,14 @@ namespace Baseline.Filesystem.Tests.FileManagerTests
         [Fact]
         public async Task It_Throws_An_Exception_If_The_Request_Was_Null()
         {
-            Func<Task> func = async () => await FileManager.DeleteAsync(null);
+            Func<Task> func = async () => await FileManager.ReadAsStringAsync(null);
             await func.Should().ThrowExactlyAsync<ArgumentNullException>();
         }
 
         [Fact]
         public async Task It_Throws_An_Exception_If_The_Path_For_The_Request_Was_Null()
         {
-            Func<Task> func = async () => await FileManager.DeleteAsync(new DeleteFileRequest());
+            Func<Task> func = async () => await FileManager.ReadAsStringAsync(new ReadFileAsStringRequest());
             await func.Should().ThrowExactlyAsync<ArgumentNullException>();
         }
 
@@ -37,19 +37,25 @@ namespace Baseline.Filesystem.Tests.FileManagerTests
         public async Task It_Throws_An_Exception_If_The_Path_Was_Obviously_Intended_As_A_Directory()
         {
             var path = "/users/Foo/bar/Destiny/XYZ/BARTINO/".AsBaselineFilesystemPath();
-            
-            Func<Task> func = async () => await FileManager.DeleteAsync(new DeleteFileRequest { FilePath = path });
+
+            Func<Task> func = async () =>
+                await FileManager.ReadAsStringAsync(new ReadFileAsStringRequest { FilePath = path }
+            );
             await func.Should().ThrowExactlyAsync<PathIsADirectoryException>();
         }
         
         [Fact]
-        public async Task It_Invokes_The_Matching_Adapters_Delete_File_Method()
+        public async Task It_Invokes_The_Matching_Adapters_ReadFileAsString_Method()
         {
             Adapter
-                .Setup(x => x.DeleteFileAsync(It.IsAny<DeleteFileRequest>(), It.IsAny<CancellationToken>()))
+                .Setup(x => x.ReadFileAsStringAsync(It.IsAny<ReadFileAsStringRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync("foo")
                 .Verifiable();
             
-            await FileManager.DeleteAsync(new DeleteFileRequest { FilePath = "a".AsBaselineFilesystemPath() });
+            var response = await FileManager.ReadAsStringAsync(
+                new ReadFileAsStringRequest { FilePath = "a".AsBaselineFilesystemPath() }
+            );
+            response.Should().Be("foo");
             
             Adapter.VerifyAll();
         }
