+++ test/Baseline.Filesystem.Tests/FileManagerTests/WriteStreamTests.cs
@@ -1,0 +1,90 @@
+using System;
+using System.IO;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using FluentAssertions;
+using Moq;
+using Xunit;
+
+namespace Baseline.Filesystem.Tests.FileManagerTests
+{
+    public class WriteStreamTests : BaseManagerUsageTest
+    {
+        [Fact]
+        public async Task It_Throws_An_Exception_If_The_Requested_Adapter_Name_Is_Not_Registered()
+        {
+            Func<Task> func = async () => await FileManager.WriteStreamAsync(
+                new WriteStreamToFileRequest
+                {
+                    FilePath = "a".AsBaselineFilesystemPath(),
+                    Stream = new MemoryStream(Encoding.UTF8.GetBytes("hello"))
+                },
+                "foo"
+            );
+            await func.Should().ThrowAsync<AdapterNotFoundException>();
+        }
+
+        [Fact]
+        public async Task It_Throws_An_Exception_If_The_Request_Was_Null()
+        {
+            Func<Task> func = async () => await FileManager.WriteStreamAsync(null);
+            await func.Should().ThrowExactlyAsync<ArgumentNullException>();
+        }
+
+        [Fact]
+        public async Task It_Throws_An_Exception_If_The_Path_For_The_Request_Was_Null()
+        {
+            Func<Task> func = async () => await FileManager.WriteStreamAsync(new WriteStreamToFileRequest
+            {
+                Stream = new MemoryStream(Encoding.UTF8.GetBytes("hello"))
+            });
+            await func.Should().ThrowExactlyAsync<ArgumentNullException>();
+        }
+
+        [Fact]
+        public async Task It_Throws_An_Exception_If_The_Stream_For_The_Request_Was_Null()
+        {
+            Func<Task> func = async () => await FileManager.WriteStreamAsync(
+                new WriteStreamToFileRequest { FilePath = "a".AsBaselineFilesystemPath() }
+            );
+            await func.Should().ThrowExactlyAsync<ArgumentNullException>();
+        }
+
+        [Fact]
+        public async Task It_Throws_An_Exception_If_The_Path_Was_Obviously_Intended_As_A_Directory()
+        {
+            var path = "/users/Foo/bar/Destiny/XYZ/BARTINO/".AsBaselineFilesystemPath();
+            
+            Func<Task> func = async () => await FileManager.WriteStreamAsync(
+                new WriteStreamToFileRequest
+                {
+                    FilePath = path,
+                    Stream = new MemoryStream(Encoding.UTF8.GetBytes("hello"))
+                }
+            );
+            await func.Should().ThrowExactlyAsync<PathIsADirectoryException>();
+        }
+        
+        [Fact]
+        public async Task It_Invokes_The_Matching_Adapters_WriteStreamToFile_Method()
+        {
+            Adapter
+                .Setup(x => x.WriteStreamToFileAsync(
+                    It.IsAny<WriteStreamToFileRequest>(), 
+                    It.IsAny<CancellationToken>())
+                )
+                .Verifiable();
+            
+            await FileManager.WriteStreamAsync(
+                new WriteStreamToFileRequest
+                {
+                    FilePath = "a".AsBaselineFilesystemPath(),
+                    Stream = new MemoryStream(Encoding.UTF8.GetBytes("hello"))
+                }
+            );
+            
+            Adapter.VerifyAll();
+        }
+    }
+}
